Fix file name issues
