+++ tests/integ_tests/core/semi_structured_vespa_index/test_schemas/semi_structured_vespa_index_schema_multiple_lexical_tensor_and_string_array_fields.sd
@@ -323,6 +323,16 @@ schema marqo__test_00semi_00structured_00schema {
             expression: modify(lexical_score(), query(marqo__mult_weights_lexical), query(marqo__add_weights_lexical))
         }
     }
+
+    rank-profile hybrid_bm25_second_phase_modifiers inherits base_rank_profile {
+        first-phase {
+            expression: lexical_score()
+        }
+        second-phase {
+            expression: modify(firstPhase, query(marqo__mult_weights_lexical), query(marqo__add_weights_lexical))
+        }
+    }
+
     rank-profile embedding_similarity inherits base_rank_profile {
         first-phase {
             expression: modify(embedding_score(), query(marqo__mult_weights_tensor), query(marqo__add_weights_tensor))
