+++ tests/unit_tests/marqo/core/search/test_hybrid_search.py
@@ -1,25 +1,26 @@
 from unittest import TestCase
 from unittest.mock import Mock, patch, PropertyMock
-from pydantic.v1 import ValidationError
+
 import semver
 
+from marqo.config import Config
 from marqo.core import constants
 from marqo.core.exceptions import UnsupportedFeatureError
-from marqo.core.models.marqo_query import MarqoHybridQuery
-from marqo.core.models.score_modifier import ScoreModifier, ScoreModifierType
-from marqo.core.models.hybrid_parameters import (
-    HybridParameters, RankingMethod, RetrievalMethod
-)
 from marqo.core.models.facets_parameters import (
     FacetsParameters, FieldFacetsConfiguration
 )
-from marqo.core.search.hybrid_search import HybridSearch
+from marqo.core.models.hybrid_parameters import (
+    HybridParameters, RankingMethod, RetrievalMethod
+)
 from marqo.core.models.marqo_index import SemiStructuredMarqoIndex, StructuredMarqoIndex
+from marqo.core.models.score_modifier import ScoreModifier, ScoreModifierType
+from marqo.core.search.hybrid_search import HybridSearch
 from marqo.core.semi_structured_vespa_index.semi_structured_vespa_index import SemiStructuredVespaIndex
+from marqo.tensor_search.enums import EnvVars
 from marqo.tensor_search.models.api_models import ScoreModifierLists, CustomVectorQuery
 from marqo.tensor_search.models.recency_parameters import RecencyParameters
 from marqo.tensor_search.models.search import SearchContext, SearchContextDocuments, SearchContextTensor
-from marqo.config import Config
+from marqo.tensor_search.utils import read_env_vars_and_defaults_ints
 
 
 class TestHybridSearch(TestCase):
@@ -338,6 +339,78 @@ class TestHybridSearch(TestCase):
         self.assertEqual(context.tensor[1].vector, [0.5, 0.6, 0.7])  # Appended tensor
         self.assertEqual(context.tensor[1].weight, 1)
 
+    @patch('marqo.core.search.hybrid_search.vespa_index_factory')
+    @patch('marqo.core.search.hybrid_search.run_vectorise_pipeline')
+    @patch('marqo.core.search.hybrid_search.utils.parse_lexical_query')
+    @patch('marqo.core.search.hybrid_search.gather_documents_from_response')
+    @patch('marqo.core.search.hybrid_search.RequestMetricsStore')
+    def test_search_custom_vector_query_with_capped_total_hits(
+            self, mock_metrics, mock_gather_docs, mock_parse_lexical,
+            mock_vectorise, mock_vespa_factory
+    ):
+        """Test CustomVectorQuery handling when context.tensor already exists (append scenario)."""
+
+        # Setup mocks
+        config = Mock(spec=Config)
+        config.vespa_client = Mock()
+        mock_response = Mock()
+        mock_response.root.coverage.coverage = 100
+        mock_response.root.coverage.degraded = None
+        config.vespa_client.query.return_value = mock_response
+
+        # Mock marqo_index
+        marqo_index = Mock(spec=SemiStructuredMarqoIndex)
+        marqo_index.name = "test_index"
+        marqo_index.parsed_marqo_version.return_value = semver.VersionInfo.parse("2.21.0")
+        marqo_index.model = Mock()
+        marqo_index.model.get_text_query_prefix.return_value = ""
+
+        # Mock vespa_index
+        mock_vespa_index = Mock(spec=SemiStructuredVespaIndex)
+        mock_vespa_query = {"query": "test"}
+        mock_vespa_index.to_vespa_query.return_value = mock_vespa_query
+        mock_vespa_factory.return_value = mock_vespa_index
+
+        # Mock vectorisation pipeline
+        mock_vectorise.return_value = {0: [0.1, 0.2, 0.3]}
+
+        # Mock lexical query parsing
+        mock_parse_lexical.return_value = (["required"], ["optional"])
+
+        # Mock metrics store
+        mock_metrics_instance = Mock()
+        mock_metrics.for_request.return_value = mock_metrics_instance
+        mock_metrics_instance.start.return_value = None
+        mock_metrics_instance.stop.return_value = 100.0
+        mock_context_manager = Mock()
+        mock_context_manager.__enter__ = Mock(return_value=None)
+        mock_context_manager.__exit__ = Mock(return_value=None)
+        mock_metrics_instance.time.return_value = mock_context_manager
+
+        # Mock gather_documents_from_response
+        mock_gather_docs.return_value = {
+            "hits": [{"_id": "1", "doc": {"field": "value"}}]
+        }
+        mock_vespa_index.gather_facets_from_response.return_value = {"totalHits": 20_000}
+
+        # Execute the search
+        hybrid_search = HybridSearch()
+        result = hybrid_search.search(
+            config=config,
+            marqo_index=marqo_index,
+            query="test",
+            track_total_hits=True,
+            hybrid_parameters=HybridParameters()
+        )
+
+        # Verify the search executed successfully
+        self.assertIsNotNone(result)
+        capped_total_hits = read_env_vars_and_defaults_ints(EnvVars.MARQO_MAX_RETRIEVABLE_DOCS)
+        self.assertEqual(
+            capped_total_hits, result["totalHits"],
+            f"The total hits should be capped to the env var value {capped_total_hits}, "
+            f"but got {result['totalHits']}"
+        )
 
 class TestRecencyValidation(TestCase):
     """Tests for recency scoring validation in HybridSearch."""
@@ -725,3 +798,104 @@ class TestRecencyValidation(TestCase):
         # The validation should pass (no error raised)
         # This test verifies that the validation ONLY fails when grow_from is provided
         self.assertIsNone(recency_params.grow_from)
+
+    @patch('marqo.core.search.hybrid_search.RequestMetricsStore')
+    def test_second_phase_modifier_on_structured_index_raises_error(self, mock_metrics):
+        """Test that secondPhaseModifier on structured index raises UnsupportedFeatureError."""
+        self._setup_metrics_mock(mock_metrics)
+
+        # Create a mock structured index
+        marqo_index = Mock(spec=StructuredMarqoIndex)
+        marqo_index.name = "test_structured_index"
+        marqo_index.model = Mock()
+        marqo_index.model.get_text_query_prefix.return_value = ""
+        marqo_index.parsed_marqo_version.return_value = semver.VersionInfo.parse("2.25.0")
+
+        config = Mock(spec=Config)
+
+        hybrid_parameters = HybridParameters(
+            secondPhaseModifier=True
+        )
+
+        hybrid_search = HybridSearch()
+        with self.assertRaises(UnsupportedFeatureError) as ctx:
+            hybrid_search.search(
+                config=config,
+                marqo_index=marqo_index,
+                query="test",
+                hybrid_parameters=hybrid_parameters
+            )
+
+        self.assertIn("secondPhaseModifier", str(ctx.exception))
+        self.assertIn("unstructured", str(ctx.exception).lower())
+
+    @patch('marqo.core.search.hybrid_search.RequestMetricsStore')
+    def test_second_phase_modifier_on_old_schema_version_raises_error(self, mock_metrics):
+        """Test that secondPhaseModifier on old schema version raises UnsupportedFeatureError."""
+        self._setup_metrics_mock(mock_metrics)
+
+        # Create a mock semi-structured index with old schema version
+        marqo_index = Mock(spec=SemiStructuredMarqoIndex)
+        marqo_index.name = "test_index"
+        marqo_index.schema_template_version = "2.24.0"  # Old version
+        marqo_index.marqo_version = "2.24.0"
+        marqo_index.model = Mock()
+        marqo_index.model.get_text_query_prefix.return_value = ""
+        marqo_index.parsed_marqo_version.return_value = semver.VersionInfo.parse("2.24.0")
+        # Does NOT support second phase lexical score modifiers
+        type(marqo_index).index_supports_second_phase_lexical_score_modifiers = PropertyMock(return_value=False)
+        marqo_index.collapse_fields = None
+
+        config = Mock(spec=Config)
+
+        hybrid_parameters = HybridParameters(
+            secondPhaseModifier=True
+        )
+
+        hybrid_search = HybridSearch()
+        with self.assertRaises(UnsupportedFeatureError) as ctx:
+            hybrid_search.search(
+                config=config,
+                marqo_index=marqo_index,
+                query="test",
+                hybrid_parameters=hybrid_parameters
+            )
+
+        self.assertIn("secondPhaseModifier", str(ctx.exception))
+        self.assertIn(str(constants.MARQO_SECOND_PHASE_LEXICAL_SCORE_MODIFIERS_MINIMUM_VERSION), str(ctx.exception))
+
+    @patch('marqo.core.search.hybrid_search.RequestMetricsStore')
+    def test_second_phase_modifier_with_collapse_fields_raises_error(self, mock_metrics):
+        """Test that secondPhaseModifier with collapse fields raises UnsupportedFeatureError."""
+        self._setup_metrics_mock(mock_metrics)
+
+        # Create a mock semi-structured index with collapse fields
+        marqo_index = Mock(spec=SemiStructuredMarqoIndex)
+        marqo_index.name = "test_index"
+        marqo_index.schema_template_version = "2.25.0"  # New version
+        marqo_index.marqo_version = "2.25.0"
+        marqo_index.model = Mock()
+        marqo_index.model.get_text_query_prefix.return_value = ""
+        marqo_index.parsed_marqo_version.return_value = semver.VersionInfo.parse("2.25.0")
+        # Supports second phase lexical score modifiers
+        type(marqo_index).index_supports_second_phase_lexical_score_modifiers = PropertyMock(return_value=True)
+        # Has collapse fields
+        marqo_index.collapse_fields = [Mock(name="category")]
+
+        config = Mock(spec=Config)
+
+        hybrid_parameters = HybridParameters(
+            secondPhaseModifier=True
+        )
+
+        hybrid_search = HybridSearch()
+        with self.assertRaises(UnsupportedFeatureError) as ctx:
+            hybrid_search.search(
+                config=config,
+                marqo_index=marqo_index,
+                query="test",
+                hybrid_parameters=hybrid_parameters
+            )
+
+        self.assertIn("secondPhaseModifier", str(ctx.exception))
+        self.assertIn("collapse", str(ctx.exception).lower())
